Refactor scenarioButtons filenames #265
