+++ visualization/app/codeCharta/ui/scenarioButtons/scenarioButtons.component.ts
@@ -6,7 +6,7 @@ import {ScenarioService, Scenario} from "../../core/scenario/scenario.service";
 import $ from "jquery";
 import {Settings, SettingsService, SettingsServiceSubscriber} from "../../core/settings/settings.service";
 import {DataModel, DataService, DataServiceSubscriber} from "../../core/data/data.service";
-import "./scenarioDropDown.scss";
+import "./scenarioDropDown.component.scss";
 
 export class ScenarioButtonsController implements TooltipServiceSubscriber, DataServiceSubscriber, SettingsServiceSubscriber {
 
@@ -75,12 +75,12 @@ export class ScenarioButtonsController implements TooltipServiceSubscriber, Data
 
 export const scenarioButtonsComponent = {
     selector: "scenarioButtonsComponent",
-    template: require("./scenarioButtons.html"),
+    template: require("./scenarioButtons.component.html"),
     controller: ScenarioButtonsController
 };
 export const scenarioDropDownComponent = {
     selector: "scenarioDropDownComponent",
-    template: require("./scenarioDropDown.html"),
+    template: require("./scenarioDropDown.component.html"),
     controller: ScenarioButtonsController
 };
