+++ visualization/app/codeCharta/ui/scenarioButtons/scenarioButtons.ts
@@ -1,11 +1,10 @@
 import angular from "angular";
-
+import "../../core/scenario/scenario.module";
 import "../../core/tooltip/tooltip.module";
-import {scenarioButtonsComponent, scenarioDropDownComponent} from "./scenarioButtonsComponent";
+import {scenarioButtonsComponent, scenarioDropDownComponent} from "./scenarioButtons.component";
 
-angular.module("app.codeCharta.ui.scenarioButtons",["app.codeCharta.core.scenario", "app.codeCharta.core.tooltip"]);
-
-angular.module("app.codeCharta.ui.scenarioButtons").component(
+angular.module("app.codeCharta.ui.scenarioButtons",["app.codeCharta.core.scenario", "app.codeCharta.core.tooltip"])
+.component(
     scenarioButtonsComponent.selector,
     scenarioButtonsComponent
 ).component(
