+++ visualization/app/codeCharta/ui/scenarioButtons/scenarioDropDown.component.scss
